style(reports): ajuste estilo de card y titulo reportes
