--- src/models/type-report.ts
@@ -1,7 +1,0 @@
-export interface TypeReport {
-    id: number;
-    createdAt: Date;
-    deletedAt: null;
-    description: string;
-    title: string;
-}
