+++ src/pages/reports/components/ReportList.tsx
@@ -5,7 +5,7 @@ import { ReportPlaceHolder } from "./ReportPlaceHolder";
 export function ReportList() {
     const { data, loading, error, itemsPerPage, totalItems, page, setPage } = useReportes();
     return (
-        <section>
+        <section className="mt-5">
             {error && <p className="bg-placeholder text-body rounded p-4 text-center uppercase">Ocurrio un error al mostrar los reportes</p>}
             {loading && (
                 <div className="grid grid-cols-1 gap-4 sm:grid-cols-3 md:grid-cols-none">
@@ -14,24 +14,35 @@ export function ReportList() {
                     ))}
                 </div>
             )}
+
             {!loading && data.length > 0 && (
                 <div className="flex flex-col gap-6">
-                    <div className="grid grid-cols-1 gap-4 md:grid-cols-1  lg:grid-cols-3 xl:grid-cols-3 ">
-                        {data
-                            .filter(reporte => reporte.state !== 'REJECTED' && reporte.state !== 'APPROVED')
-                            .map((reporte) => (
-                                <ReportCard key={reporte.id} report={reporte} wrapperExtraClasses='py-4' />
-                            ))}
-                    </div>
+                    {
+                        data.filter(reporte => reporte.state === 'PENDING').length <= 0 ? (
+                            <div className="flex flex-col gap-4">
+                                <p className="bg-placeholder text-body rounded p-4 text-center uppercase">No hay reportes pendientes</p>
+                            </div>
+                        ) : (<>
+                            <div className="flex flex-col gap-4">
+                                <div className="grid grid-cols-1 gap-4 md:grid-cols-1  lg:grid-cols-3 xl:grid-cols-3 ">
+                                    {data
+                                        .filter(reporte => reporte.state === 'PENDING')
+                                        .map((reporte) => (
+                                            <ReportCard key={reporte.id} report={reporte} wrapperExtraClasses='py-4' />
+                                        ))}
+                                </div>
+                            </div>
+                        </>)
+                    }
                     <Pagination
                         itemsPerPage={itemsPerPage}
                         totalItems={totalItems}
                         page={page}
                         setPage={setPage} />
-                </div>
+                </div >
             )}
 
-        </section>
+        </section >
 
     )
 }
